backend/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohappyeyeballs==2.4.4
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.10.1
astral==3.2
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.4
cached-property==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
Click==7.0
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==45.0.3
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.6.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.19.1
elastic-transport==8.17.1
elasticsearch==8.18.1
email_validator==2.2.0
entrypoints==0.3
exceptiongroup==1.2.2
fake-useragent==2.0.0
fastapi==0.115.12
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
graphql-core==3.2.6
greenlet==3.1.1
h11==0.14.0
h3==4.2.1
httpcore==1.0.7
httplib2==0.14.0
httptools==0.1.2
httpx==0.28.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib_resources==6.4.5
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==6.1.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
passlib==1.7.4
pexpect==4.6.0
pillow==10.4.0
propcache==0.2.0
psycopg2-binary==2.9.10
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.20.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-telegram-bot==21.6
pytz==2025.1
PyYAML==5.3.1
reportlab==4.4.1
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.9.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.4
soupsieve==2.7
SQLAlchemy==2.0.28
ssh-import-id==5.10
starlette==0.44.0
strawberry-graphql==0.255.0
systemd-python==234
texttable==1.6.2
timezonefinder==6.5.8
Twisted==18.9.0
typing_extensions==4.12.2
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.11.3
uvloop==0.21.0
wadllib==1.3.3
websocket-client==0.53.0
websockets==8.1
wsproto==0.15.0
yarl==1.15.2
zipp==3.20.2
zope.interface==4.7.1
